# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/dashboard/resumo": {
      "get": {
        "operationId": "DashboardController_getResumo",
        "summary": "Total de fazendas e de hectares cadastrados",
        "parameters": [
          {
            "name": "estado",
            "required": false,
            "in": "query",
            "description": "Sigla do estado (UF) para recortar o resultado. Omitido = todos os estados.",
            "schema": {
              "example": "MG",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResumoDto"
                }
              }
            }
          },
          "401": {
            "description": "Token JWT ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "dashboard"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/dashboard/por-estado": {
      "get": {
        "operationId": "DashboardController_getDistribuicaoPorEstado",
        "summary": "Quantidade de propriedades agrupadas por estado (UF)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DashboardPorEstadoDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token JWT ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "dashboard"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/dashboard/por-cultura": {
      "get": {
        "operationId": "DashboardController_getDistribuicaoPorCultura",
        "summary": "Quantidade de plantios agrupados por cultura",
        "parameters": [
          {
            "name": "estado",
            "required": false,
            "in": "query",
            "description": "Sigla do estado (UF) para recortar o resultado. Omitido = todos os estados.",
            "schema": {
              "example": "MG",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DashboardPorCulturaDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token JWT ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "dashboard"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/dashboard/uso-do-solo": {
      "get": {
        "operationId": "DashboardController_getUsoDoSolo",
        "summary": "Soma de área agricultável x área de vegetação",
        "parameters": [
          {
            "name": "estado",
            "required": false,
            "in": "query",
            "description": "Sigla do estado (UF) para recortar o resultado. Omitido = todos os estados.",
            "schema": {
              "example": "MG",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardUsoDoSoloDto"
                }
              }
            }
          },
          "401": {
            "description": "Token JWT ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "dashboard"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Brain Agriculture API",
    "description": "API REST para gestão de cadastro de produtores rurais, suas propriedades e os plantios por safra/cultura, incluindo o dashboard de indicadores.\n\nAutenticação: faça login em `POST /auth/login` e use o `accessToken` retornado como `Authorization: Bearer <token>` nas demais requisições (botão \"Authorize\" acima).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "timestamp": {
            "type": "string",
            "example": "2026-08-26T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/produtores"
          },
          "message": {
            "type": "string",
            "example": "Dados inválidos"
          },
          "errors": {
            "example": [
              "nome should not be empty"
            ],
            "description": "Presente quando o erro vem do ValidationPipe (um item por campo inválido).",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "statusCode",
          "timestamp",
          "path",
          "message"
        ]
      },
...
      "DashboardResumoDto": {
        "type": "object",
        "properties": {
          "totalFazendas": {
            "type": "number",
            "example": 42
          },
          "totalHectares": {
            "type": "number",
            "example": 15320.5,
            "description": "Soma da área total (ha) de todas as propriedades"
          }
        },
        "required": [
          "totalFazendas",
          "totalHectares"
        ]
      },
      "DashboardPorEstadoDto": {
        "type": "object",
        "properties": {
          "estado": {
            "type": "string",
            "example": "MG",
            "description": "Sigla do estado (UF)"
          },
          "total": {
            "type": "number",
            "example": 12,
            "description": "Quantidade de propriedades cadastradas nesse estado"
          }
        },
        "required": [
          "estado",
          "total"
        ]
      },
      "DashboardPorCulturaDto": {
        "type": "object",
        "properties": {
          "cultura": {
            "type": "string",
            "example": "Soja"
          },
          "total": {
            "type": "number",
            "example": 8,
            "description": "Quantidade de plantios registrados para essa cultura"
          }
        },
        "required": [
          "cultura",
          "total"
        ]
      },
      "DashboardUsoDoSoloDto": {
        "type": "object",
        "properties": {
          "areaAgricultavel": {
            "type": "number",
            "example": 9800.3,
            "description": "Soma da área agricultável (ha) de todas as propriedades"
          },
          "areaVegetacao": {
            "type": "number",
            "example": 5520.2,
            "description": "Soma da área de vegetação (ha) de todas as propriedades"
          }
        },
        "required": [
          "areaAgricultavel",
          "areaVegetacao"
        ]
      },
...
    }
  }
}